Rock-paper-scissors-lizard-Spock
This program plays a game of Rock-paper-scissors-lizard-Spock with the computer. The player's choices are predefined and the computer's choices are generated. A winner is chosen based on the rules of the game.

This code was written in CodeSkulptor (http://www.codeskulptor.org/) and takes advantage of the built-in functionality on that platform.