Guess the number
In this guessing game, you're given a range and asked to guess the secret number which can range from 0 to 99, or 999. In either case, you're allowed a total number of guesses based on the size of the range. (Hint: use the Binary Search algorithm to do your guessing).

The output is between the console and the GUI. Buttons allow you to change the range and text input allows you to enter your guess.

This code was written in CodeSkulptor (http://www.codeskulptor.org/) and takes advantage of the functionality available on that platform.