Create Stopwatch: The Game
This game creates a stopwatch with three buttons: start, stop, and reset. The goal is to stop the watch on a whole second (ie., with no fractional seconds). A running score at the top keeps track of successful stops and total stops.

This code was written in CodeSkulptor (http://www.codeskulptor.org/) and takes advantage of the functionality available on that platform.